requirements.txt
certifi==2016.2.28
click==6.7
enum34==1.1.6
Flask==0.12.2
futures==3.1.1
imageio==2.4.1
itsdangerous==0.24
Jinja2==2.9.6
json-tricks==3.12.2
kafka-python==1.4.3
MarkupSafe==1.0
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
Werkzeug==0.12.2
wincertstore==0.2
